Fixed segmentation fault
Fixed CURL overwriting std::string - forgot about linux x64 data model

+++ src/HttpClient.h
@@ -12,7 +12,7 @@ namespace Http
 	struct Response
 	{
 		int error;
-		int status;
+		long status;
 		std::string body;
 	};
